requirements.txt
streamlit
langchain-groq
python-dotenv
pdf2image
langchain-google-genai
